requirements.txt
playwright>=1.45.0
python-dotenv>=1.0.0
requests>=2.31.0
PyYAML>=6.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
